Check all non-success states when attempting to clear a previous snapshot
A snapshot can be in four states: IN_PROGRESS, SUCCESS, FAILED, or PARTIAL. To
distinguish between the case when a snapshot has never been attempted and the
case when a snapshot has been attempted but didn't succeed, we were checking to
see if the snapshot state was FAILED (since in that case, we need to delete the
bad snapshot metadata in order to retry). We recently saw a snapshot hanging
around in the PARTIAL state because of an error from AWS, so this patch
changes the backup code to look for any non-success state to determine whether
we should clear the snapshot metadata.