Remove plugin check from docs
All our ES instances have the plugin by default and this README is
obviously meant for deployment on Aptible. If someone deploys off of
Aptible, then they'll get an error running the cron about having to
install the plugin, which is probably enough (bonus: it'll also be *the
right plugin depending on the ES version*).